Started a general refactor based upon golint
I was playing around with this library and after looking around, saw some
new things that I wanted to add. Before doing this, I ran a golint to see
if there were a few things I could tidy up first instead.

A lot of this is simply removing `else` from conditional statements whereby
it would `return` before getting there anyway. Due to this, the compiler would
then complain about variables that have been declared but not used. I've noticed
the following pattern being used a fair amount:

```go
if something, err := something.DoSomething(); err != nil {
    return err
}
```

and due to this the variables need to be declared (above a statement like this)
before they can be used in this way, i.e.

```go
var (
    something aType
    err       error
)
```

which could also be written in this way:

```go
something, err := something.DoSomething()
if err != nil {
    return err
}
```

In this case, as to not change too much in the way you've done this, I've
gone with the former for now. As mentioned, these things are mostly linter
related issues, with a few other small refactors, including comments on
exported types and funcs. After running `golint` on this codebase since,
there are much fewer errors and most of the ones that still remain are simply
telling you to comment things that are exported.

There are some other improvements I'd like to make here, but for now this is a
start.

Thanks for putting together a go client library for this. It's been fun to play
around with it!