Started a general refactor based upon golint (#8)
* Started a general refactor based upon golint

I was playing around with this library and after looking around, saw some
new things that I wanted to add. Before doing this, I ran a golint to see
if there were a few things I could tidy up first instead.

A lot of this is simply removing `else` from conditional statements whereby
it would `return` before getting there anyway. Due to this, the compiler would
then complain about variables that have been declared but not used. I've noticed
the following pattern being used a fair amount:

```go
if something, err := something.DoSomething(); err != nil {
    return err
}
```

and due to this the variables need to be declared (above a statement like this)
before they can be used in this way, i.e.

```go
var (
    something aType
    err       error
)
```

which could also be written in this way:

```go
something, err := something.DoSomething()
if err != nil {
    return err
}
```

In this case, as to not change too much in the way you've done this, I've
gone with the former for now. As mentioned, these things are mostly linter
related issues, with a few other small refactors, including comments on
exported types and funcs. After running `golint` on this codebase since,
there are much fewer errors and most of the ones that still remain are simply
telling you to comment things that are exported.

There are some other improvements I'd like to make here, but for now this is a
start.

Thanks for putting together a go client library for this. It's been fun to play
around with it!

* Reverted if blocks where func scope vars were previously created

As discussed in #8, any code that involved creating new vars in the
func scope to get past the linter, has been reverted.

* Changing const names to use CamelCase over full caps

i.e. OAUTH2_CLIENT_ID > Oauth2ClientID

* Rearranged Grids type, changed case for Oauth2 consts, using Nodes type

Just a few of the small changes listed in the PR. This doesn't include changes
regarding `if else` issues just yet.

* Modified isSSL func nodeInitial and nodeStatus

So for these I have gone with your suggestions:

for isSSL, it's now simply:

```go
if configURL, err := url.Parse(config.URL); err != nil {
    return false
} else {
    return configURL.Scheme == "https"
}
```

and for the other two conditionals, we're now checking the special case
first.